Add ability to find simulator by name and runtime
In Xcode 9.2 the simulators window name changed from the form
"iPhone X - iOS 11.1" to "iPhone X- 11.2". This change breaks the tunnel
from finding the correct simulator window because it performs a complete
string comparison.
To fix this we add the ability to specify the device name and device
runtime in the request parameters (the change is backward compatible, so
old clients can continue to use the server version). With the new
specified parameters we can identify all the simulator window that
follow that contains the device name and the runtime version.